requirements.txt
aiogram==3.22.0
aiohttp==3.12.15
asyncpg==0.30.0
SQLAlchemy==2.0.43
python-dotenv==1.1.1
openai==1.107.2
yookassa==3.6.0
APScheduler==3.11.0
greenlet==3.2.4
